ml/requirements.txt
fastapi
uvicorn
deepface
tf-keras
numpy
pillow
open_clip_torch
transformers